[Add] Attr class분리 및 멤버변수 명 변경
